[engine][server] Refactoring. Moved DealsTest/TestUtils/AbstractDealsTest to engine.
