+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/services/InactivityServiceTest.java
@@ -26,6 +26,7 @@ import com.codenjoy.dojo.services.multiplayer.Sweeper;
 import com.codenjoy.dojo.services.settings.CheckBox;
 import com.codenjoy.dojo.services.settings.EditBox;
 import com.codenjoy.dojo.services.settings.Settings;
+import com.codenjoy.dojo.utils.test.AbstractDealsTest;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -40,6 +41,7 @@ import static org.mockito.Mockito.*;
 public class InactivityServiceTest extends AbstractDealsTest {
 
     private InactivityService inactivity;
+    private TimeService timeService;
 
     @Before
     public void setup() {
