+++ CodingDojo/server/src/test/java/com/codenjoy/dojo/services/dao/ActionLoggerTest.java
@@ -32,6 +32,7 @@ import com.codenjoy.dojo.services.multiplayer.MultiplayerType;
 import com.codenjoy.dojo.services.multiplayer.Spreader;
 import com.codenjoy.dojo.services.printer.BoardReader;
 import com.codenjoy.dojo.services.room.RoomService;
+import com.codenjoy.dojo.utils.test.TestUtils;
 import lombok.SneakyThrows;
 import org.junit.After;
 import org.junit.Before;
